refactor: moved user folder
