+++ api/src/main/java/lab/en2b/quizapi/auth/AuthService.java
@@ -5,8 +5,8 @@ import lab.en2b.quizapi.auth.dtos.*;
 import lab.en2b.quizapi.auth.jwt.JwtUtils;
 import lab.en2b.quizapi.commons.exceptions.TokenRefreshException;
 import lab.en2b.quizapi.auth.dtos.RefreshTokenResponseDto;
-import lab.en2b.quizapi.user.User;
-import lab.en2b.quizapi.user.UserService;
+import lab.en2b.quizapi.commons.user.User;
+import lab.en2b.quizapi.commons.user.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.authentication.AuthenticationManager;
