+++ api/src/main/java/lab/en2b/quizapi/auth/config/SecurityConfig.java
@@ -1,6 +1,6 @@
 package lab.en2b.quizapi.auth.config;
 
-import lab.en2b.quizapi.user.UserService;
+import lab.en2b.quizapi.commons.user.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
