+++ api/src/main/java/lab/en2b/quizapi/auth/jwt/JwtAuthFilter.java
@@ -3,7 +3,7 @@ import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import lab.en2b.quizapi.user.UserService;
+import lab.en2b.quizapi.commons.user.UserService;
 import lombok.NonNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
